Move all food api tests to one file to reduce duplicated setup blocks add sad path for /api/v1/foods/:badID
